requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pyodbc==4.0.30

